requirements.txt
django==3.2.13
mysqlclient==2.1.1
django-simpleui==2024.3.25
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.14.0
django-filter==22.1
requests==2.28.2
requests_toolbelt==0.10.1
jsonpath==0.82
pymysql==1.0.2
rsa==4.9
faker==17.6.0
redis==4.5.4
celery==5.2.7
django-celery-beat==2.5.0
pyyaml==6.0
uritemplate==4.1.1
coreapi==2.3.3
drf-yasg==1.21.5
gunicorn==20.1.0
supervisor==4.2.5
eventlet==0.36.0